+++ vm/src/system/program/tests/mod.rs
@@ -32,12 +32,12 @@ fn test_flatten_fromslice_roundtrip() {
     assert_eq!(num_cols, flattened.len());
 }
 
-fn interaction_test(program: Program<BabyBear>, execution: Vec<usize>) {
+fn interaction_test(program: Program<BabyBear>, execution: Vec<u32>) {
     let instructions = program.instructions();
     let mut chip = ProgramChip::new(program);
     let mut execution_frequencies = vec![0; instructions.len()];
     for pc in execution {
-        execution_frequencies[pc] += 1;
+        execution_frequencies[pc as usize] += 1;
         chip.get_instruction(pc).unwrap();
     }
     let air = chip.air.clone();
@@ -137,7 +137,7 @@ fn test_program_negative() {
     let mut chip = ProgramChip::new(program);
     let execution_frequencies = vec![1; instructions.len()];
     for pc in 0..instructions.len() {
-        chip.get_instruction(pc).unwrap();
+        chip.get_instruction(pc as u32).unwrap();
     }
     let air = chip.air.clone();
     let cached_trace = chip.generate_cached_trace();
